refactor(sandbox): run registered bash/writeFile/readFile tools on Sprites
Reconcile the merged PR4 registration with the Sprites driver. Split the
flag-gated, default-OFF tools so the provider SDK never loads in the factory's
tests:

- sandbox-tools.ts is now the provider-agnostic factory only (schemas +
  context resolution + call-time gate + delegation); no DB, no backing-provider
  SDK import.
- sandbox-tools-runtime.ts holds the production wiring (DB-backed session store,
  Fly Sprites driver, quota, audit, actor resolver) and the gate wiring, and
  exports buildSandboxTools.
- ai-tools.ts imports buildSandboxTools from the runtime module; registration
  stays flag-gated and default-OFF (PR4 behaviour preserved, just on Sprites).
