+++ apps/web/src/lib/ai/core/ai-tools.ts
@@ -14,7 +14,7 @@ import { calendarReadTools } from '../tools/calendar-read-tools';
 import { calendarWriteTools } from '../tools/calendar-write-tools';
 import { channelTools } from '../tools/channel-tools';
 import { workflowTools } from '../tools/workflow-tools';
-import { buildSandboxTools } from '../tools/sandbox-tools';
+import { buildSandboxTools } from '../tools/sandbox-tools-runtime';
 import { CORE_TOOL_NAMES } from './stub-tools';
 
 const baseTools = {
